rename SourceAsset to FileSource
