+++ crates/turbopack-core/src/file_source.rs
