+++ crates/turbopack-core/src/lib.rs
@@ -13,6 +13,7 @@ pub mod compile_time_info;
 pub mod context;
 pub mod environment;
 pub mod error;
+pub mod file_source;
 pub mod ident;
 pub mod introspect;
 pub mod issue;
@@ -24,7 +25,6 @@ pub mod reference;
 pub mod reference_type;
 pub mod resolve;
 pub mod server_fs;
-pub mod source_asset;
 pub mod source_map;
 pub mod source_pos;
 pub mod source_transform;
