+++ crates/turbopack-node/src/transforms/webpack.rs
@@ -7,10 +7,10 @@ use turbo_tasks_fs::{json::parse_json_with_source_context, File, FileContent};
 use turbopack_core::{
     asset::{Asset, AssetContent, AssetContentVc, AssetVc},
     context::{AssetContext, AssetContextVc},
+    file_source::FileSourceVc,
     ident::AssetIdentVc,
     module::ModuleVc,
     reference_type::{InnerAssetsVc, ReferenceType},
-    source_asset::FileSourceVc,
     source_transform::{SourceTransform, SourceTransformVc},
     virtual_asset::VirtualAssetVc,
 };
